core/requirements.txt
# Core.
Flask
werkzeug
gunicorn

# Testing and static analysis.
pytest
pytest-cov
flake8

# CLI.
Click

# Data and workers.
psycopg2
Flask-SQLAlchemy
SQLAlchemy
sqlalchemy_utils

# Utils.
Flask-WTF
WTForms-Components
fake-factory
requests
pyqrcode
pypng

# Extensions.
ipfshttpclient
ipfsapi
pybase62
pytz